+++ tests/test_process.py
@@ -807,6 +807,25 @@ class EnvironmentTest(CoverageTest):
         actual = self.run_command(f"python {cov_main} run run_me.py")
         self.assert_tryexecfile_output(expected, actual)
 
+    def test_pythonsafepath(self) -> None:
+        with open(TRY_EXECFILE) as f:
+            self.make_file("run_me.py", f.read())
+        self.set_environ("PYTHONSAFEPATH", "1")
+        expected = self.run_command("python run_me.py")
+        actual = self.run_command("coverage run run_me.py")
+        self.assert_tryexecfile_output(expected, actual)
+
+    @pytest.mark.skipif(env.PYVERSION < (3, 11), reason="PYTHONSAFEPATH is new in 3.11")
+    def test_pythonsafepath_dashm(self) -> None:
+        with open(TRY_EXECFILE) as f:
+            self.make_file("with_main/__main__.py", f.read())
+
+        self.set_environ("PYTHONSAFEPATH", "1")
+        expected = self.run_command("python -m with_main")
+        actual = self.run_command("coverage run -m with_main")
+        assert re.search("No module named '?with_main'?", actual)
+        assert re.search("No module named '?with_main'?", expected)
+
     def test_coverage_custom_script(self) -> None:
         # https://github.com/nedbat/coveragepy/issues/678
         # If sys.path[0] isn't the Python default, then coverage.py won't