+++ CHANGES.rst
@@ -23,7 +23,8 @@ upgrading your version of coverage.py.
 Unreleased
 ----------
 
-Nothing yet.
+- Fix: the PYTHONSAFEPATH environment variable new in Python 3.11 is properly
+  supported, closing `issue 1696`_.  Thanks, `Philipp A. <pull 1700_>`_.
 
 
 .. scriv-start-here
@@ -281,6 +282,8 @@ Version 7.3.3 — 2023-12-14
 
   - ``dataop2`` writes the full data being added to CoverageData objects.
 
+.. _issue 1696: https://github.com/nedbat/coveragepy/issues/1696
+.. _pull 1700: https://github.com/nedbat/coveragepy/pull/1700
 .. _issue 684: https://github.com/nedbat/coveragepy/issues/684
 .. _pull 1705: https://github.com/nedbat/coveragepy/pull/1705
 .. _issue 1709: https://github.com/nedbat/coveragepy/issues/1709